% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosumvar_factor.R
\name{ar.weighted}
\alias{ar.weighted}
\title{Fit an autoregressive time series model to the data, using weighted estimators}
\usage{
ar.weighted(x, cps = NULL, weight.method = c("linear", "exp", "robust"), ...)
}
\arguments{
\item{x}{matrix of data with series as columns}

\item{cps}{integer (vector) of estimated change points}

\item{weight.method}{String of weight method to use}

\item{...}{further arguments to \code{ar}}
}
\value{
\code{ar.weighted} object, see \link[stats]{ar}
}
\description{
Fit an autoregressive time series model to the data, using weighted estimators
}
\examples{
fm <- fnets::fnets.factor.model(t(panel$panel), fm.restricted = TRUE, q = 2)
mod <- ar.weighted(fm$factors, cps = 100)
predict(mod, fm$factors, n.ahead = 5)
}
