% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{fvar.sim}
\alias{fvar.sim}
\title{Simulate from a piecewise stationary factor model with VAR dynamics}
\usage{
fvar.sim(
  n,
  p = 100,
  q = 2,
  order = 1,
  cps = c(),
  signal = 0.7,
  error.dist = c("normal", "t", "garch"),
  P1 = NULL,
  Q1 = NULL,
  df = 3
)
}
\arguments{
\item{n}{sample size}

\item{p}{number of series}

\item{q}{factor number}

\item{order}{VAR order}

\item{cps}{change points}

\item{signal}{size of parameters}

\item{error.dist}{error distribution for VAR and idiosyncratic errors, one of \code{"normal", "t", "garch"}}

\item{P1}{see \link[mosumvar]{VAR.sim}}

\item{Q1}{see \link[mosumvar]{VAR.sim}}

\item{df}{see \link[mosumvar]{VAR.sim}}
}
\value{
List containing
\itemize{
    \item{\code{x}}{ observed series}
    \item{\code{f}}{ factor series}
    \item{\code{e}}{ error series}
    \item{\code{lam}}{ factor loadings}
    \item{\code{cps}}{ change points}
}
}
\description{
Simulate from a piecewise stationary factor model with VAR dynamics
}
\examples{
data <- fvar.sim(500, cps = 200)
}
