% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosumvar_factor.R
\name{get.weights}
\alias{get.weights}
\title{Get regression weights}
\usage{
get.weights(n, cps, weight.method)
}
\arguments{
\item{n}{sample size}

\item{cps}{integer (vector) of estimated change points}

\item{weight.method}{String of weight method to use}
}
\value{
weight vector
}
\description{
Get regression weights
}
\keyword{internal}
