% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosumvar_factor.R
\name{lm.weighted}
\alias{lm.weighted}
\title{Fit a linear model to the data, using weighted estimators}
\usage{
lm.weighted(
  y,
  x,
  intercept = FALSE,
  cps = NULL,
  weight.method = c("linear", "exp", "robust"),
  ...
)
}
\arguments{
\item{y}{vector of responses}

\item{x}{matrix of data with series as columns}

\item{intercept}{add intercept in regression}

\item{cps}{integer (vector) of estimated change points}

\item{weight.method}{String of weight method to use}

\item{...}{further arguments to \code{lm}}
}
\value{
\code{lm} object, see \link[stats]{lm}
}
\description{
Fit a linear model to the data, using weighted estimators
}
\examples{
fm <- fnets::fnets.factor.model(t(panel$panel), fm.restricted = TRUE, q = 2)
lm.weighted(panel$gdp, fm$factors, cps = 100)
}
