% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosumvar_factor.R
\name{mosumfvar}
\alias{mosumfvar}
\title{Segment data under a factor model with VAR dynamics}
\usage{
mosumfvar(
  x,
  center = TRUE,
  q = c("ic", "er"),
  order = NULL,
  G = NULL,
  method = c("Score", "Wald"),
  estim = c("C", "H"),
  var.estim = c("Local", "Global"),
  alpha = 0.05,
  criterion = c("eps", "eta"),
  nu = 0.25,
  do.bootstrap = FALSE,
  n.bootstrap = 1000,
  thresh = NULL,
  do.plot = TRUE,
  algo = c("mosumvar", "univ", "ms"),
  rm.cross.terms = TRUE,
  global.resids = TRUE
)
}
\arguments{
\item{x}{matrix of data with series as columns}

\item{center}{whether to de-mean the input \code{x}}

\item{q}{Either the number of factors or a string specifying the factor number selection method; possible values are:
\itemize{
   \item{\code{"ic"}}{ information criteria-based methods of Alessi, Barigozzi & Capasso (2010) when \code{fm.restricted = TRUE} or Hallin and Liška (2007) when \code{fm.restricted = FALSE} modifying Bai and Ng (2002)}
   \item{\code{"er"}}{ eigenvalue ratio of Ahn and Horenstein (2013)}
}
see \link[fnets]{factor.number}.}

\item{order}{integer VAR model order}

\item{G}{integer MOSUM bandwidth (or vector, if \code{algo = "ms"}); see reference for default}

\item{method}{detector, one of \code{"Wald", "Score"}}

\item{estim}{estimator method, one of \code{"C", "H"}}

\item{var.estim}{variance estimator method, one of \code{"Local", "Global"}}

\item{alpha}{Numeric significance level}

\item{criterion}{location procedure, one of \code{"eps", "eta"}}

\item{nu}{Numeric location procedure hyperparameter}

\item{do.bootstrap}{Boolean, determine threshold via bootstrap method}

\item{n.bootstrap}{Integer; number of bootstrap replicates}

\item{thresh}{rejection threshold; see reference for default}

\item{do.plot}{Boolean, return plot}

\item{algo}{which algorithm to use, one of \code{"mosumvar","univ","ms"}}

\item{rm.cross.terms}{Boolean, remove cross terms when \code{univ = TRUE}}

\item{global.resids}{Boolean, use residuals from full VAR model when \code{univ = TRUE}}
}
\value{
List of class \code{mosumfvar}, containing
\itemize{
  \item{\code{seg}}{ resulting segmentation, a \code{mosumvar} object. See \link[mosumvar]{mosumvar}.}
  \item{\code{fm}}{ fitted factor model, a \code{fm} object. See \link[fnets]{fnets.factor.model}.}
}
}
\description{
Segment data under a factor model with VAR dynamics
}
\examples{
mosumfvar(panel$panel, order = 1,  method = "Score", q = 2)
}
